move auth to User (#5)
