+++ src/user/routes.rs
@@ -1,4 +1,3 @@
-use crate::auth;
 use crate::post::Post;
 use crate::user::{User, UserPublic, UserRequest};
 use actix_web::{delete, get, post, put, web, HttpResponse, Responder};
@@ -111,7 +110,7 @@ async fn find_posts(id: web::Path<Uuid>, db_pool: web::Data<PgPool>) -> impl Res
 
 #[post("/users/login")]
 async fn login(credentials: BasicAuth, db_pool: web::Data<PgPool>) -> impl Responder {
-    let result = auth::basic_auth_validator(credentials, db_pool.get_ref()).await;
+    let result = User::basic_auth(credentials, db_pool.get_ref()).await;
     match result {
         Ok(user) => HttpResponse::Ok().json(UserPublic::from(user)),
         Err(err) => {
